Changed code to use a single FlmDbSystem object.  Fixed utilities so that they build using FTK.
git-svn-id: http://svn.code.sf.net/p/flaim/code/trunk@405 0109f412-320b-0410-ab79-c3e0c5ffbbe6

+++ xflaim/java/jni/jdb.cpp
@@ -311,7 +311,7 @@ Exit:
 	}
 	
 	pEnv->ReleaseStringCritical( sElementName, pszElementName);
-	return( (jlong)uiNameId);
+	return( (jint)uiNameId);
 }
 
 /****************************************************************************
